tracking/requirements.txt
fastapi==0.58.1
uvicorn==0.11.5
pymongo==3.10.1
PyJWT==1.7.1
